Simulate different doses/intervals in a dose/interval grid

Simulate different doses/intervals in a dose/interval grid

Usage

simulate_dose_interval(
  value,
  grid_type = "dose",
  dose_update,
  regimen,
  md,
  pta,
  target_design,
  model,
  t_obs,
  omega,
  obs,
  ruv,
  ...
)

Arguments

value

element of the dose/interval grid

grid_type

either dose grid or interval grid

dose_update

update dose from which dose?

regimen

PKPDsim regimen object

md

metadata object (only needed if we have to use get_quantity_from_variable() to generate target value)

pta

probability of target attainment, list with arguments type and value, also requires omega if non-NULL. If NULL, will just aim for specific conc or auc.

target_design

object specifying target design, created using function create_target_design()

model

model for simulating dose (estimation model)

t_obs

time at which observation should be calculated

omega

IIV matrix, for estimation model, for probability of target attainment target types.

obs

Value of obs as determined by dose_grid_search() (i.e. either "obs" or AUC compartment)

ruv

list specifying residual error for estimation model: list(prop = 0.1, add = 1.5), for probability of target attainment target types.

...

passed on to PKPDsim function